put deps-related stuff in to own package - ultimately every git action is tied to info in the DepInfo structs
