+++ toml_cfg/toml_cfg.go
@@ -2,20 +2,14 @@ package toml_cfg
 // vim: noet ts=4 sw=4 sr smartindent:
 
 import (
-	"errors"
     "fmt"
+	"dep"
 
 	toml "github.com/BurntSushi/toml"
 )
 
 type TomlCfg struct {
-	Deps map[string]DepInfo
-}
-
-type DepInfo struct {
-	Src   string
-	Ref   string
-	Depth string
+	Deps map[string]dep.DepInfo
 }
 
 func Read(toml_file string) (TomlCfg, error) {
@@ -43,19 +37,6 @@ func (c TomlCfg) ValidateDeps(toml_file string) error {
 		if err != nil {
 			return err
 		}
-	}
-
-	return err
-}
-
-func (d DepInfo) ValidateRef(clone_dir string) error {
-
-	var err error
-
-	if d.Ref == "" {
-		err_tmpl := "ERROR: ref not supplied for src %s in [deps.%s]\n"
-		err_msg := fmt.Sprintf(err_tmpl, d.Src, clone_dir)
-		err = errors.New(err_msg)
 	}
 
 	return err
